Move block spec to its own module
